+++ src/main/java/za/co/hpsc/web/enums/Division.java
@@ -83,7 +83,7 @@ public enum Division {
      * @return an {@code Optional} containing the matching {@code Division} if found,
      * or empty otherwise.
      */
-    public static Optional<Division> getByName(String name) {
+    public static Optional<Division> fromName(String name) {
         if ((name == null) || (name.isBlank())) {
             return Optional.empty();
         }
@@ -107,7 +107,7 @@ public enum Division {
      * @return an {@code Optional} containing the matching {@code Division} if found,
      * or empty otherwise.
      */
-    public static Optional<Division> getByAbbreviation(String abbreviation) {
+    public static Optional<Division> fromAbbreviation(String abbreviation) {
         if ((abbreviation == null) || (abbreviation.isBlank())) {
             return Optional.empty();
         }
@@ -131,7 +131,7 @@ public enum Division {
      * @return an {@code Optional} containing the matching {@code Division} if found,
      * or empty if no match is found.
      */
-    public static Optional<Division> getByAbbreviationOrName(String value) {
+    public static Optional<Division> fromAbbreviationOrName(String value) {
         if ((value == null) || (value.isBlank())) {
             return Optional.empty();
         }
@@ -156,7 +156,7 @@ public enum Division {
      * @return an {@code Optional} containing the matching {@code Division} if found,
      * or empty otherwise.
      */
-    public static Optional<Division> getByCode(Integer code) {
+    public static Optional<Division> fromCode(Integer code) {
         if ((code == null) || (code <= 0)) {
             return Optional.empty();
         }